Add USR2, QUIT, ABRT in list of handled signals
When starting unicorn inside a container using sdnotify-proxy, we need a
way to pass USR2 and QUIT signals to it. We also need a way to pass ABRT
signal to kafka-consumers when they run inside containers.

This commit adds the above to the list of signals handled by
sdnotify-proxy.

The build script is also removed as it is not useful for our purposes.